Splitting monolithic actions/index into various files. Using actions/index as a stopgap to refactoring all imports. Clearing temp variable for function in actions/class.
