+++ kolibri/plugins/facility_management/assets/src/state/actions/data.js
@@ -1,0 +1,13 @@
+import { PageNames } from '../../constants';
+
+import { _managePageTitle } from './helpers/mappers';
+import preparePage from './helpers/preparePage';
+
+export default function showDataPage(store) {
+  preparePage(store.dispatch, {
+    name: PageNames.DATA_EXPORT_PAGE,
+    title: _managePageTitle('Data'),
+    isAsync: false,
+  });
+  store.dispatch('SET_PAGE_STATE', {});
+}
