+++ kolibri/plugins/facility_management/assets/src/state/actions/helpers/displayModal.js
@@ -1,0 +1,5 @@
+export default function displayModal(store, modalName) {
+  store.dispatch('SET_MODAL', modalName);
+  store.dispatch('SET_ERROR', '');
+  store.dispatch('SET_BUSY', false);
+}
